Fix signed/unsigned warning.

+++ src/build.cc
@@ -495,7 +495,7 @@ void RealCommandRunner::Abort() {
 bool RealCommandRunner::CanRunMore() {
   size_t subproc_number =
       subprocs_.running_.size() + subprocs_.finished_.size();
-  bool can_run_more = (subproc_number < config_.parallelism
+  bool can_run_more = (subproc_number < (size_t)config_.parallelism
     && ((subprocs_.running_.empty() || config_.max_load_average <= 0.0f)
         || GetLoadAverage() < config_.max_load_average));
